# server/docs/postman-openapi-full.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zalo Clone — Chat Server API (full)",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://unlimited-feedback-crossing-academics.trycloudflare.com",
      "description": "Qua Cloudflare tunnel (gọi từ máy khác)"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local trực tiếp tại máy chủ"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "AGENT_API_KEY hoặc ADMIN_API_KEY trong `server/.env`. GET /api/media/* được miễn."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Tham số không hợp lệ / thiếu.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Thiếu hoặc sai API key (header `X-Api-Key`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mô tả lỗi (tiếng Việt)."
          },
          "code": {
            "type": "string",
            "nullable": true,
            "description": "Mã lỗi gốc từ Zalo (nếu có)."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Passthrough": {
        "type": "object",
        "additionalProperties": true,
        "description": "Object thô trả về từ Zalo (zca-js). Cấu trúc tuỳ thao tác."
      },
...
    }
  },
  "paths": {
...
    "/api/groups": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Tạo nhóm",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Tham số tạo nhóm (name, members…) theo zca-js."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/members-info": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Thông tin nhiều thành viên",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Danh sách id ngăn cách bởi dấu phẩy."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/invite": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Mời một người vào nhiều nhóm",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "groupIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "userId",
                  "groupIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/link-info": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Thông tin nhóm theo link mời",
        "parameters": [
          {
            "name": "link",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberPage",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/join-link": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Tham gia nhóm bằng link",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "link": {
                    "type": "string"
                  }
                },
                "required": [
                  "link"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/invite-box": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Danh sách lời mời vào nhóm",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/invite-box/{groupId}": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Chi tiết lời mời vào nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      },
      "delete": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Xoá lời mời vào nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "blockFutureInvite",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/invite-box/{groupId}/join": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Chấp nhận lời mời vào nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Thông tin nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      },
      "delete": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Giải tán nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/members": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Thêm thành viên",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "memberId": {
                    "type": "string"
                  }
                },
                "required": [
                  "memberId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/members/{memberId}": {
      "delete": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Xoá thành viên",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/deputies": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Thêm phó nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "memberId": {
                    "type": "string"
                  }
                },
                "required": [
                  "memberId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/deputies/{memberId}": {
      "delete": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Gỡ phó nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/owner": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Chuyển chủ nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "memberId": {
                    "type": "string"
                  }
                },
                "required": [
                  "memberId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/name": {
      "patch": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Đổi tên nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/avatar": {
      "patch": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Đổi ảnh nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatarSource": {
                    "description": "Nguồn ảnh (URL/base64…)."
                  }
                },
                "required": [
                  "avatarSource"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/settings": {
      "patch": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Cập nhật cài đặt nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/blocked": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Danh sách bị chặn trong nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      },
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Chặn thành viên",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "memberId": {
                    "type": "string"
                  }
                },
                "required": [
                  "memberId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/blocked/{memberId}": {
      "delete": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Bỏ chặn thành viên",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/pending": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Danh sách chờ duyệt vào nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/pending/review": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Duyệt/từ chối yêu cầu vào nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "members": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isApprove": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "members",
                  "isApprove"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/link/enable": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Bật link nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/link/disable": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Tắt link nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/link": {
      "get": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Chi tiết link nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/groups/{groupId}/leave": {
      "post": {
        "tags": [
          "Nhóm"
        ],
        "summary": "Rời nhóm",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "silent": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
...
  }
}